requirements.txt
Flask
pandas
scikit-learn
email-validator
dnspython
